Handle binary images with a missing uuid
Occasionally a crash report has binary images with a missing uuid field. This caused the the rest of the binary image list to be ignored, because the parser thought it had reached the last record.

This fix makes the uuid optional. Any binary image with a missing uuid will be skipped, and a warning message is emitted in verbose mode.